% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convPlot.R
\name{convPlot}
\alias{convPlot}
\title{Convergence plot for the inclusion probabilities of a Bvs (Gibbs) object.}
\usage{
convPlot(object, variables = "All")
}
\arguments{
\item{object}{An object of class \code{Bvs} obtained with \code{GibbsBvs}}

\item{variables}{A vector containing the name of the variables that would be plotted (all of them by default)}
}
\value{
\code{convPlot} returns a picture of the trace of the inclusion probabilities of the variables included in \code{variable}
}
\description{
Returns a picture of the trace of the inclusion probabilities of the potential variables
}
\examples{
\dontrun{
#Analysis of Ozone35 data

data(Ozone35)

#We use here the (Zellner) g-prior for
#regression parameters and constant prior
#over the model space
#In this Gibbs sampling scheme, we perform 10100 iterations,
#of which the first 100 are discharged (burnin) and of the remaining
#only one each 10 is kept.
#as initial model we use the Full model
Oz35.GibbsBvs<- GibbsBvs(formula= y ~ ., data=Ozone35, prior.betas="gZellner",
prior.models="Constant", n.iter=10000, init.model="Full", n.burnin=100,
time.test = FALSE)

#Convergence plot for all the variables
convPlot(Oz35.GibbsBvs)

#Convergence plot for x4 x5 and x6
convPlot(Oz35.GibbsBvs, variables = c("x4", "x5", "x6"))
}
}
\author{
Anabel Forte and Gonzalo Garcia-Donato
}
