\name{Hald}
\alias{Hald}
\docType{data}
\title{
Hald data
}
\description{
The following data relates to an engineering application that was interested in the effect of the cement composition on heat evolved during hardening (for more details, see Woods et al., 1932).}
\usage{data(Hald)}
\format{
  A data frame with 13 observations on the following 5 variables.
  \describe{
    \item{\code{y}}{Heat evolved per gram of cement (in calories)}
    \item{\code{x1}}{Amount of tricalcium aluminate}
    \item{\code{x2}}{Amount of tricalcium silicate}
    \item{\code{x3}}{Amount of tetracalcium alumino ferrite}
    \item{\code{x4}}{Amount of dicalcium silicate}
  }
}
\references{
	Woods, H., Steinour, H. and Starke, H. (1932). 
	Effect of Composition of Porland Cement on Heat Evolved During Hardening. Industrial 	and Engineering Chemistry Research, 24, 1207-1214.
}
\examples{
data(Hald)
}
\keyword{datasets}
