% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltltn.R
\name{pltltn}
\alias{pltltn}
\title{Correction for p>>n for an object of class \code{Bvs}}
\usage{
pltltn(object)
}
\arguments{
\item{object}{An object of class \code{Bvs} obtained with \code{GibbsBvs}}
}
\value{
\code{pltltn} returns a list with the following elements: 
\item{pS }{An estimation of the probability that the true model is irregular (k>n)} 
\item{postprobdim }{A corrected estimation of the posterior probabilities over the dimensions}
\item{inclprob }{A corrected estimation of the posterior inclusion probabilities}
}
\description{
In cases where p>>n and the true model is expected to be sparse, it is very unlikely that the Gibbs sampling
will sample models in the singular subset of the model space (models with k>n). Nevertheless, depending on
how large is p/n and the strenght of the signal, this part of the model space could be very influential in the
final response.
}
\details{
From an object created with GibbsBvs and prior probabilities specified as Scott-Berger, 
this function provides an estimation of the posterior probability of models with k>n which is a measure of the 
importance of these models. In summary, when this probability is large,  the sample size is not large enough to beat
such large p.
Additionally, \code{pltltn} gives corrections of the posterior inclusion probabilities and posterior probabilities 
of dimension of the true model.
}
\examples{

\dontrun{
#Analysis of
}

}
\seealso{
See 
  \code{\link[BayesVarSel]{GibbsBvs}} for creating objects of the class
  \code{Bvs}.
}
\author{
Gonzalo Garcia-Donato

Maintainer: <gonzalo.garciadonato@uclm.es>
}
