% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsBvsF.R
\name{summary.BvsF}
\alias{summary.BvsF}
\title{Summary of an object of class \code{BvsF}}
\usage{
\method{summary}{BvsF}(object, ...)
}
\arguments{
\item{object}{An object of class \code{BvsF}}

\item{...}{Additional parameters to be passed}
}
\description{
Summary of an object of class \code{BvsF}, providing inclusion probabilities and a representation of
the Median Probability Model and the Highest Posterior probability Model.
}
\examples{


}
\seealso{
See \code{\link[BayesVarSel]{Bvs}},
  \code{\link[BayesVarSel]{GibbsBvs}} for creating objects of the class
  \code{Bvs}.
}
\author{
Gonzalo Garcia-Donato and Anabel Forte

  Maintainer: <anabel.forte@uv.es>
}
