% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{Hald}
\alias{Hald}
\title{Hald data}
\format{A data frame with 13 observations on the following 5 variables.
\describe{ \item{y}{Heat evolved per gram of cement (in calories)}
\item{x1}{Amount of tricalcium aluminate} \item{x2}{Amount
of tricalcium silicate} \item{x3}{Amount of tetracalcium alumino
ferrite} \item{x4}{Amount of dicalcium silicate} }}
\usage{
Hald
}
\description{
The following data relates to an engineering application that was interested
in the effect of the cement composition on heat evolved during hardening
(for more details, see Woods et al., 1932).
}
\examples{
data(Hald)

}
\references{
Woods, H., Steinour, H. and Starke, H.
(1932)<DOI:10.1021/ie50275a002> Effect of Composition of Porland Cement on
Heat Evolved During Hardening. Industrial and Engineering Chemistry
Research, 24, 1207-1214.
}
\keyword{datasets}
