% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{OBICE}
\alias{OBICE}
\title{OBICE data}
\format{
A data frame with 1188 entries and 251 variables. Among the more relevant are
\describe{
\item{ActFisica}{:(physic activity) factor coded as 1 None; 2 less than monthly; 3 less than weekly; 4 less than 2/week; 5 at least 2/week; 9 is NA}
\item{ActivDepor}{: weekly hours devoted to sports activity (9 is NA)}
\item{Acostarse}{: does he/she eat before going to bed? (yes/no)}
\item{Bebida}{: (main dring accompanying the main meal) 1 water tap; 2 bottle water; 3 soda; 4 natural juices; 5 bottle juices; 6 Milk (and derivatives); 7 Other; (9 is NA)}
\item{PadreObeso}{: is the father obese? (0 is No; 1 is Yes)}
\item{MadreObesa}{: is the mother obese? (0 is No; 1 is Yes)}
\item{PesoNac}{: weight born (in grams) (9999 is NA)}
\item{PesoActual}{: current weight (in kilograms)}
\item{TallaNac}{: height born (in centimeters) (999 is NA)}
\item{TallaAct}{: current height (in meters)}
\item{CincoComidas}{: does he/she have regularly 5 meals per day? (0 is No; 1 is Yes)}
\item{Verduras}{: usual consumption of vegetables? (0 is No; 1 is Yes)}
\item{Fruta}{: usual consumption of fruit? (0 is No; 1 is Yes)}
\item{Edad}{: years old}
\item{Sexo}{: female (1); male (0)}
\item{LactMatern}{: breast-feeding (1 is Yes; 0 is No)}
\item{Merienda}{: Afternoon snack  (1 is Yes; 0 is No)}
\item{Chuches}{: Sweets and soft drinks weekly consumption (how many times) (9 is NA)}
\item{HorasPCDiaPond}{: daily hours playing videogames and/or in internet (weekends included) (NA is NA)}
\item{HorasTVDiaPond}{: daily hours watching TV (weekends included) (NA is NA)}
\item{HoraSuenyo}{: daily hours sleeping (99 is NA)}
}
}
\usage{
OBICE
}
\description{
Dataset corresponding to the OBICE study (Zurriaga et al 2011) where factors associated with childhood obesity
are studied. The data were collected in 2007 and 2008 through several questionnaries and 
n=1188 children were enroled in the study. It contains 155 variables.
This is a case and control study with 437 cases (obese) and 751 controls (not obese).
Purposedly the dataset is distributed without any post-processing
hence, many variables may contain unavailable observations coded in different way.
}
\examples{
data(OBICE)

}
\references{
Zurriaga, O., Perez-Panades, J., Quiles, J. , Gil, M.,
Anes, Y., Quiñones, C., Margolles, M., Lopez-Maside, A.,
Vega-Alonso, A., Miralles M. and Recent OBICE Research Group (2011)
Factors associated with childhood obesity in Spain. The OBICE
study: a case–control study based on sentinel networks. Public Health Nutrition
14(6), 1105–1113.
}
\keyword{datasets}
