% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltltn.R
\name{pltltn}
\alias{pltltn}
\title{Correction for p>>n for an object of class \code{Bvs}}
\usage{
pltltn(object)
}
\arguments{
\item{object}{An object of class \code{Bvs} obtained with \code{GibbsBvs}}
}
\description{
In cases where p>>n and the true model is expected to be sparse, it is very unlikely that the Gibbs sampling
will not sample models in the singular set of the model space (models with k>n). Nevertheless, depending on
how large is p/n and the strenght of the signal, this part of the model space could be very influential in the
final response.
}
\details{
This function provides an estimation of the posterior probability of models with k>n which is a measure of the 
importance of these models. In summary, if this probability is large, implies that the n is not large enough to beat
such large p.
Additionally, \code{pltltn} gives corrections of the posterior inclusion probabilities and posterior probabilities 
of dimension of the true model.
}
\examples{

\dontrun{
#Analysis of Crime Data
#load data
data(UScrime)

#Default arguments are Robust prior for the regression parameters
#and constant prior over the model space
#Here we keep the 1000 most probable models a posteriori:
crime.Bvs<- Bvs(formula= y ~ ., data=UScrime, n.keep=1000)

#A look at the results:
summary(crime.Bvs)
}

}
\seealso{
See 
  \code{\link[BayesVarSel]{GibbsBvs}} for creating objects of the class
  \code{Bvs}.
}
\author{
Gonzalo Garcia-Donato

  Maintainer: <gonzalo.garciadonato@uclm.es>
}
